chore: move RPC to an env var, fix some imports
