+++ typescript/optics-deploy/.env.example
@@ -1,3 +1,8 @@
 ALFAJORES_DEPLOYER_KEY=
+ALFAJORES_RPC=
+
 KOVAN_DEPLOYER_KEY=
+KOVAN_RPC=
+
 RINKEBY_DEPLOYER_KEY=
+RINKEBY_RPC=
